+++ src/fetch-url.js
@@ -1,21 +1,14 @@
 import { config as configStore } from './stores';
+import { isMapboxUrl, normalizeMapboxUrl } from './mapbox-urls';
 
 let mapboxGlAccessToken;
 configStore.subscribe(value => ({ mapboxGlAccessToken } = value));
-
-const isMapboxUrl = url => {
-  if (typeof url !== 'string') return false;
-  const hasMapboxFormat =
-    url.startsWith('mapbox://styles/') && url.split('/').length === 5;
-  return hasMapboxFormat;
-};
 
 const fetchUrl = async url => {
   const urlIsMapbox = isMapboxUrl(url);
   let nextUrl = url;
   if (urlIsMapbox) {
-    const [, , , userName, styleId] = nextUrl.split('/');
-    nextUrl = `https://api.mapbox.com/styles/v1/${userName}/${styleId}?title=true&access_token=${mapboxGlAccessToken}`;
+    nextUrl = normalizeMapboxUrl(url, mapboxGlAccessToken);
   }
   let response;
   try {